# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The pi-cloud-monitor RESTful API Documents.",
        "title": "pi-cloud-monitor RESTful API",
        "contact": {
            "name": "wukong",
            "email": "wukong@pplabs.org"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "47.114.74.103:20000",
    "basePath": "/v1",
    "paths": {
        "/user/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户名和密码",
                        "name": "ReqUserLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ReqUserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RspUserLogin"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误"
                    },
                    "500": {
                        "description": "内部错误"
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ReqUserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 5
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 5
                }
            }
        },
        "handler.RspUserLogin": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}